Use the term 'seal' to describe the uncompressed header
This 'seal' contains the DSA signature, but also a few more
things. This use of 'seal' is probably not in line with the
use of the term in the relevant UIC docs, but it allows me
to distinguish the full uncompressed header from the DSA
signature, which is only part of that header. It also allows
me to distinguish the uncompressed header (henceforth called
'seal') from the compressed header (U_HEAD).